cifssrv: Add support for compounding requests
Compounding:
If we look at the command set for the new SMB2 protocol, they are all simple
operations part of the reduced command set. The SMB1.0 protocol had some
complex commands and subcommands that combined a set of simple operations
as required in specific scenarios.

One of the important changes in SMB2 is the ability to send an arbitrary set
of commands in a single request (single network round trip). This is called
compounding and it can be use to mimic the old complex operations in SMB1
without the added complexity of a larger command set.

For instance, an old SMB1 RENAME command can be replaced by a single request
in SMB2 that combines three commands: CREATE (which can create a new file or
open an existing file), SET_INFO and CLOSE. The same can be done for many
other complex SMB1 commands and subcommands like LOCK_AND_READ and
WRITE_AND_UNLOCK.

This compounding ability in SMB2 is very flexible and the chain of commands
can be unrelated (executed separately, potentially in parallel) or related
(executed in sequence, with the output of one command available to the next).
The responses can also be compounded or sent separately.
This new compounding feature in SMB2 can be used to perform a specific task
in less time due to the reduced number of network round trips.

Note: Currently only Windows client utilize this Componding feature and
requests for compounding requests.
From linux-variants (cifs) clients - there are no compound requests yet.

Signed-off-by: Namjae Jeon <namjae.jeon@samsung.com>