cifssrv: initialization of request buffer and error handling in smb_trans2
Request buffers are not properly initialized (not zeroed out)
which is resulting in GARBAGE data being read from the SMB REQUEST,
at some points - it shows reading the FREED data with values like 0x6b6b
(in this case showed TRANS2 sub command to be this value).
So, get zeroed allocation for the SMB requested buffers
and also
in SMB_TRANS2 - setupcount should be atleast 1 and greater.
So, the checking corrected to avoid unnecessary WARN_ON from base.smb testcase

Signed-off-by: Namjae Jeon <namjae.jeon@samsung.com>