cifssrv: addition of new dialects to IS_SMB2 family
Few features like oplock/leases are checked based upon the IS_SMB2
family category, while the support for newer dialects like
SMB2.x/SMB3.0/SMB3.02
SMB3.11 is added correctly but the same is not reflected in the IS_SMB2
macro.
As a result, few features support was missing and causing performance drop.
Adding the dialects to the IS_SMB2 macros.

Signed-off-by: Mayank Singh <mayank.singh@samsung.com>
Reviewed-by: Amit Sahrawat <a.sahrawat@samsung.com>
Signed-off-by: Namjae Jeon <namjae.jeon@protocolfreedom.org>