cifssrv: fix random kernel oops due to netlink socket creation failure
As the netlink init error was not captured in the CIFSSRV startup
routine, it resulted in undefined netlink socket.
So, when initiating command - it tried to send request to te error socket,
resulting in kernel oops.
Since, this is also part initialization routines - the server should
not start if the netlink init part fails, adding error handling here.

Signed-off-by: Amit Sahrawat <a.sahrawat@samsung.com>
Signed-off-by: Namjae Jeon <namjae.jeon@protocolfreedom.org>