cifssrv: fix crash in smb_grant_oplock with windows
server thread was exited after 120 seconds, but there was still
an open file. When share is mounted next time again on windows
it finds that there is already open file and lease was granted.
Next it tries to find lease owner of last opened file. In this
process it tries to access server address which was already freed
due to inactivity for 120 seconds.

In order to fix this issue, do not exit server thread if there is
any open file.

Previously it was crashing like below backtrace:

Unable to handle kernel paging request at virtual address 6b6b6b6f
...
[<c0292bcc>] (smb_grant_oplock) from [<c029b824>] (smb2_open+0x1cc4/0x2160)
[<c0299b60>] (smb2_open) from [<c0276784>] (handle_smb_work+0x378/0x60c)
[<c027640c>] (handle_smb_work) from [<c0042578>] (process_one_work+0x140/0x458)
[<c0042438>] (process_one_work) from [<c00428e4>] (worker_thread+0x54/0x4b8)
[<c0042890>] (worker_thread) from [<c00483a4>] (kthread+0xec/0x104)
[<c00482b8>] (kthread) from [<c000ff98>] (ret_from_fork+0x14/0x3c)

Signed-off-by: Vivek Trivedi <t.vivek@samsung.com>
Signed-off-by: Namjae Jeon <namjae.jeon@protocolfreedom.org>