cifssrv: fix memleak in smb_vfs_listxattr
Allocated memory is tracked through the passed pointer in the
argument, Once the allocation is success - the memory
should be immediately assigned. Because, the caller is checking
the pointer before freeing.

The error conditions can be used only for changing the error
value to -E2BIG

This fixes memory leaks from the the filesystem specific xattr handlers
returning error's other than ERANGE.

Signed-off-by: Amit Sahrawat <a.sahrawat@samsung.com>
Signed-off-by: Namjae Jeon <namjae.jeon@protocolfreedom.org>