cifsd: put connection request on a diet
Our usage of kmem cache for request buffers is completely off
target, because kmem cache is good for structures of a regular
size. Whilst out connection request can be either small, or
large, or super large (vmalloc()).

Small requests are requests of up to MAX_CIFS_SMALL_BUFFER_SIZE
bytes. But we allocate both small and large buffers, regardless
of the PDU size.

Large requests are in the range from MAX_CIFS_SMALL_BUFFER_SIZE
to SMBMaxBufSize + max_hdr_size. Which is around 65K.

This means that even a 1K request will use a 65K kmem cache
buffer - a tremendous waste of memory.

Anything larger than SMBMaxBufSize + max_hdr_size goes into
a gian buffer, which is named ->wbuf, to completely confuse
anyone. And there is a dynamic switch of buffers, which also
involves memcpy(), and so on.

When we queue smb work, we used to do a very hard to follow
games with the buffers - because we basically have 3 of them
in connection - small, large and super large. And then we had
a bunch of bool flags to decide which one to assign to smb
work and which one to free later.

The truth is:
- we don't need to use kmem cache here
- we don't need to have 3 buffers
- we don't need to switch between them
- we don't need to have 3 bool flags
- we know the PDU size, so we can allocate just one buffer
  of appropriate size.

And this is what this patch is doing. It gets rid of small and
large request buffer pools, it get rid of 3 pointers and bool
flags. So connection now uses just one buffer, exacetly of the
PDU size.

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>