cifsd: do not check for cifsd PID
Drop the cifsd PID check. This is not a proper way of ensuring
that user space is up and running. PID may exists, but the app
may deadlock, for instance.

If we need heart beats - then we need to have proper heart
beats on the protocol level (IPC), say NL heart beat request
every 10 seconds [e.g. ping].

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>