cifsd: rename cifsd_mfp to cifsd_inode
cifsd_mfp name is misleading. It stands for (I'm not sure it's
even documented anywhere) for "master file pointer".

First, "master" suggests that we have "slave", which is not true.
Second, it's not a "file pointer", actually, but "inode".

Rename it to cifsd_inode, so the name will represent what it
actually is - cifsd extra data which we have associated with every
vfs inode.

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>