cifsd: rework IPC heartbeat frame
heartbeat frame cannot be a sleeping request; because this will
fail when a thread which sent heartbeat request is the same thread
which reads IPC messages.

We can use a simpler scheme: we can't send IPC frame to a dead PID.
So if we successfully sent heartbeat frame to ->snd_portid then
it means that it's still there, listening on the netlink socket.
We will fail to send a hertbeat to a dead ->snd_portid.

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>