cifsd: attempt to simplify handler code
So this patch tries to re-work the TCP handler code. It's very complex,
it have goto up, down, up goto form the middle, and so on. This is way
to hard to follow. For example, suppose that both checks return TRUE
in the example below:

===

chained:
       rc = check_conn_state(work);
       if (rc)
               goto send;

[...]

send:
       if (is_chained_smb2_message(work))
               goto chained;
===

We will just ping-pong; that's hard to debug. So this patch tries
to simplify the code by factoring out some parts of it and by
introduction of some clear return values and actions.

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>