cifsd: rollback to unlock codes for waiting task
This is a patch to rollback codes.

In case of a notify task that is currently being implemented, it will be
set to asynchronous and wait for any file change events. Therefore, the
locking should be released before cmds->proc(), so that the next commands
from clients can be processed.

int smb2_notify(struct cifsd_work *cifsd_work)
{
    ...

    setup_async_work();
    smb2_send_interim_resp();

    // waiting for notification here
    cifsd_file_notify_wait();
    ...
}

Signed-off-by: Yunjae Lim <yunjae.lim@samsung.com>