cifsd: rework extract_sharename()
For every treename extract_sharename() does
	while (strchr(treename, '\\'))
		strsep(&treename, "\\");

So, for example, for treename '\\localhost\IPC$', it scans treename
from the start, finds the first \\, invokes strsep(), which scans
treename from the start and updates the treename pointer with the
position of the first \\. Then it repeats whe whole thing for the
second \\ and for the last \\, until it finds IPC$ part.

We don't have to do that. We can call strchr() just once, to find
the last \\, and be done with the whole thing. No loops, no strsep().

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>