cifsd: move file_table back to session
Volatile (8 bytes): A file handle that can be changed when
an open is reconnected after being lost on a disconnect, as
specified in section 3.3.5.9.7. The server MUST return this
file handle as part of an SMB2 CREATE Response (section 2.2.14).
This value MUST NOT change unless a reconnection is performed.
This value MUST be unique for all volatile handles within the
scope of a session.

+ further API cleanup

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>