cifsd: remove fp f_state
It doesn't look like we need fp ->f_state and FP_NEW/FP_FREEING anymore.
fp is either valid and is on the ci list of fps, or it's invalid and
thus is NOT on the list.

All we need is to be careful with ci m_lock.

add/remove: 0/0 grow/shrink: 1/4 up/down: 43/-111 (-68)
Function                                     old     new   delta
__cifsd_lookup_fd                            111     154     +43
cifsd_open_fd                                712     703      -9
__cifsd_close_fd.isra                        811     782     -29
cifsd_close_inode_fds                        471     439     -32
cifsd_smb_check_shared_mode                  546     505     -41

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>