cifsd: optimize cifsd_vfs_empty_dir()
We don't need to allocate a PAGE_SIZE buffer for cifsd_vfs_empty_dir(),
the only thing we care about is the number of ->dentries. So we don't
even need to call cifsd_fill_dirent() and to bother with all the
memcpy() and so on.

Intoroduce a simple cifsd_count_dir_entries() which simply increments
->dirent_count.

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>