cifsd: optimize cifsd_vfs_lookup_in_dir()
We don't need a PAGE_SIZE buffer in order to lookup
a file within the dir. In fact, we don't need any extra
memory at all - we almost have everything we need.

Stop allocating __get_free_page() for cifsd_vfs_lookup_in_dir()
and avoid calling cifsd_fill_dirent(). That function is not what
we need here. We also can stop iterating readdir_data.dirent.

What we need is to extend readdir_data with a private pointer,
through which we pass the pointer to the file name which we lookup
and, at the same, to the buffer to which we should copy the actual
dentry name. So now iterate_dir() ->actor can access all the required
data and we can stop doing a ton of memcpy() and so on.

add/remove: 1/0 grow/shrink: 0/1 up/down: 90/-353 (-263)
Function                                     old     new   delta
__caseless_lookup                              -      90     +90
cifsd_vfs_kern_path                          733     380    -353
Total: Before=183806, After=183543, chg -0.14%

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>