cifsd: rework cifsd_vfs_readdir_name()
I really have no idea why cifsd_vfs_readdir_name() is returning
a pointer to extra memory.

We always have a proper ->name for de and dir_path. Then
cifsd_vfs_readdir_name() allocates extra buffer in order to
create a full name dir_path + '/' + de->name. OK, so far so good.

But then... It moves de->name to the beginning of that temp buffer
and NULL-s out the rest. So basically it return back the de->name,
which is stored in extra memory.

Which we then need to carefully kfree().

	de = buffer + offset; // here we have proper de->name set
	...
	de->name = cifsd_vfs_readdir_name(...de);
	...
	kfree(de->name);

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>