cifsd: add name_len member to struct cifsd_dir_info
When we process dentries smb2_query_dir()->__query_dir()
we know the length of the dentry name, yet for some reason
when we call smb2_populate_readdir_entry() we forget it
and pass PATH_MAX to convname_updatenextoffset() as the
size of d_info->name.

Add ->name_len member to d_info, set it appropriately and
pass right name length value to convname_updatenextoffset().

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>