cifsd: rework smb2 query_dir
Recent smb2 query_dir changes turned out to be wrong and ugly,
basically because we landed a recursion there:

 query_dir		[cifsd]
  iterate_dir		[vfs]
   .readdir		[fs]
    ctx->actor
     __query_dir	[cifsd]
      kern_path		[vfs]
       .lookup		[fs]

It turned out that not all file systems handle that recursion
nicely. Namely vfat, which doesn't handle nested nor parallel
operations. vfat takes the SB mutex in .readdir and calls ctx
->actor from under that SB mutex, then cifsd invokes kern_path()
from ->actor() which performs dcache lookup. If the requested
path is in dcache then we are fine, but if we need to read it
from the disk, we end up in file system's .lookup callback.
At this point vfat deadlocks, because it attempts to take the
same SB mutex in .lookup, which we already hold in .readdir:

  iterate_dir		[vfs]
   .readdir		[fs]
   mutex_lock(SB)
    ctx->actor
     __query_dir	[cifsd]
      kern_path		[vfs]
       .lookup		[fs]
       mutex_lock(SB)	[<deadlock>]

Unfortunately, there is no easy way out of this, and we need
to perform qeury dir in two steps - iterate dir first, then
do dentry lookups to fill up dentry's attrs and xattrs.

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>