cifsd: use lookup_one_len()
cifsd_vfs_readdir_name() is a bit heave. It takes the dirpath
string and file name string, allocates a temp buffer, combines
dirpath and file name strings to perform a full path lookup:
	kern_path(dirpath + '/' + filename)

We don't need that. At this point we already have dir path
and dir dentry and a filename. So all we need to do is to
lookup_one_len(), IOW to lookup just filename.

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>