cifsd: always zero-initialize ksmbd_dir_info in smb1pdu
For most errors the find_next() and find_first() implementations jump
to an err_out label. From there kfree(d_info.smb1_name) is called
unconditionally. Use memset() to zero-out the struct ksmbd_dir_info
d_info variable at the beginning of the functions so when any error
occurs d_info.smb1_name will either be NULL or have a valid pointer in
it.

This fixes a crash which was observed in kmem_cache_free_bulk() -
presumably because the code was trying to free memory from an address
which was just garbage (non-initialized) data from the stack.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>