cifsd: prevent ksmbd from sending duplicate oplock break notifications
Sending duplicate oplock break notifications can make trouble
, which is solved in the commit 6e39685b359d ("ksmbd: avoid to send
duplicate oplock break notifications")'. But this can happen
again in the following situations.

1) Right after a notification is sent and opinfo->pending_break
bit lock is released, A break request for the same oplock level
succeeds to hold opinfo->pending_break bit lock and a duplicate
noticfication is sent.

2) A notification is pending in cifs, But File id is reused and
a notification for the id is sent to cifs.

This patch solves the above.

Signed-off-by: Hyunchul Lee <hyc.lee@gmail.com>