ksmbd: smb1: Fix FID allocation range
Per MS-CIFS 2.2.1.6.1:
  The value 0xFFFF MUST NOT be used as a valid FID. All other possible
  values for FID, including zero (0x0000) are valid. The value 0xFFFF is
  used to specify all FIDs or no FID, depending upon the context in
  which it is used.

This fixes an issue that can be reproduced with the following steps:
 - on the server, create a directory with > 64K files:
   for i in $(seq 1 66000); do echo "file" > ${i}; done
 - on the client, mount the share with vers=1 and access each file:
   find path/to/share -type f -exec grep -n 'fakepattern' {} \;

This effectively makes a series of Read_AndX/Close requests. Since the
server allocated FID=0xFFFF, the close request on that FID will
invalidate all other FIDs, which results in the client getting "Bad file
descriptor" errors for all remaining files.

Signed-off-by: Marios Makassikis <mmakassikis@freebox.fr>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>