ksmbd: smb1: fix null dereference in smb_session_disconnect
smb_session_disconnect() calls ksmbd_session_destroy() which checks
for null before freeing the user, so remove the call.

The null-deref can be triggered by a client sending a Session Setup
request with a NtLmNegotiate payload (therefore allocating and
registering a session) followed by a Session Disconnect request.
Because users are allocated on NtLmAuthenticate, sess->user is null.

Signed-off-by: Marios Makassikis <mmakassikis@freebox.fr>
Signed-off-by: Namjae Jeon <linkinjeon@kernel.org>